+++ database/migrations/2023_04_27_000740_create_donors_table.php
@@ -22,7 +22,7 @@ class CreateDonorsTable extends Migration
             $table->string('email')->unique();
             $table->string('gender', 8);
             $table->string('chro_dis');
-            $table->integer('blood_type');
+            $table->unsignedBigInteger('blood_type');
             $table->timestamps();
 
             $table->foreign('blood_type')->references('ref')->on('blood_categories')->onDelete('cascade');
